Refactor complete
