+++ rust/sleeper_core/src/lib.rs
@@ -1,5 +1,10 @@
-//! The `internal` module contains the internal functionality and error conditions
-//! to actually implement the compaction library.
+//! The `compaction` crate implements all the core functionality for running Sleeper
+//! Parquet data compaction in Rust. We provide a C library interface wrapper which
+//! will serve as the interface from Java code in Sleeper. We are careful to adhere to C style
+//! conventions here such as libc error codes.
+//!
+//! We have an internal "details" module that encapsulates the internal workings. All the
+//! public API should be in this module.
 /*
  * Copyright 2022-2025 Crown Copyright
  *
@@ -15,13 +20,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+mod datafusion;
+
 use aws_config::Region;
 use aws_credential_types::Credentials;
 use color_eyre::eyre::{Result, bail, eyre};
 use object_store::aws::AmazonS3Builder;
 use objectstore_ext::s3::{ObjectStoreFactory, config_for_s3_module, default_creds_store};
-use std::{collections::HashMap, path::PathBuf};
+use std::collections::HashMap;
 use url::Url;
+
+pub use datafusion::sketch::{DataSketchVariant, deserialise_sketches};
 
 /// Type safe variant for Sleeper partition boundary
 #[derive(Debug, Copy, Clone)]
@@ -125,7 +134,7 @@ pub struct CompactionResult {
 /// # use url::Url;
 /// # use aws_types::region::Region;
 /// # use std::collections::HashMap;
-/// # use crate::sleeper_df::{run_compaction, CompactionInput, PartitionBound, ColRange};
+/// # use crate::sleeper_core::{run_compaction, CompactionInput, PartitionBound, ColRange};
 /// let mut compaction_input = CompactionInput::default();
 /// compaction_input.input_files = vec![Url::parse("file:///path/to/file1.parquet").unwrap()];
 /// compaction_input.output_file = Url::parse("file:///path/to/output").unwrap();
@@ -201,23 +210,14 @@ async fn create_object_store_factory(
     ObjectStoreFactory::new(s3_config)
 }
 
-fn to_s3_config(aws_config: &AwsConfig) -> AmazonS3Builder {
+/// Create an [`AmazonS3Builder`] from the given configuration object.
+///
+/// Credentials are extracted from the given configuration object.
+#[must_use]
+pub fn to_s3_config(aws_config: &AwsConfig) -> AmazonS3Builder {
     let creds = Credentials::from_keys(&aws_config.access_key, &aws_config.secret_key, None);
     let region = Region::new(String::from(&aws_config.region));
     config_for_s3_module(&creds, &region)
         .with_endpoint(&aws_config.endpoint)
         .with_allow_http(aws_config.allow_http)
-}
-
-/// Creates a file path suitable for writing sketches to.
-///
-#[must_use]
-pub fn create_sketch_path(output_path: &Url) -> Url {
-    let mut res = output_path.clone();
-    res.set_path(
-        &PathBuf::from(output_path.path())
-            .with_extension("sketches")
-            .to_string_lossy(),
-    );
-    res
 }
