+++ rust/sleeper_core/tests/compaction_helpers.rs
@@ -1,5 +1,18 @@
-use std::{collections::HashMap, fs::File, sync::Arc};
-
+/*
+ * Copyright 2022-2025 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 use arrow::{
     array::{Array, ArrayRef, Int32Array, RecordBatch},
     datatypes::{DataType, Field, Schema},
@@ -13,8 +26,9 @@ use datafusion::parquet::{
     basic::{Compression, ZstdLevel},
     file::properties::WriterProperties,
 };
-use sleeper_df::{ColRange, PartitionBound};
-use sleeper_df::{DataSketchVariant, deserialise_sketches};
+use sleeper_core::{ColRange, PartitionBound};
+use sleeper_core::{DataSketchVariant, deserialise_sketches};
+use std::{collections::HashMap, fs::File, sync::Arc};
 use tempfile::TempDir;
 use url::Url;
