+++ rust/sleeper_core/tests/compaction_test.rs
@@ -1,11 +1,25 @@
+/*
+ * Copyright 2022-2025 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 mod compaction_helpers;
-
-use std::{collections::HashMap, sync::Arc};
 
 use arrow::datatypes::{DataType, Field, Schema};
 use color_eyre::eyre::Error;
 use compaction_helpers::*;
-use sleeper_df::{CompactionInput, run_compaction};
+use sleeper_core::{CompactionInput, run_compaction};
+use std::{collections::HashMap, sync::Arc};
 use tempfile::tempdir;
 use test_log::test;
