+++ rust/sleeper_df/Cargo.toml
@@ -13,7 +13,7 @@
 # limitations under the License.
 [package]
 name = "sleeper_df"
-description = "This is the native library for Sleeper that implements all the core data compaction and sketches generation functionality."
+description = "This is the native library for Sleeper that implements all of the Foreign Function Interface logic."
 keywords = ["sleeper"]
 categories = ["science"]
 version = { workspace = true }
@@ -31,31 +31,27 @@ test = true
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-arrow = { workspace = true, features = ["prettyprint"] }
-aggregator_udfs = { path = "../aggregator_udfs" }
-aws-config = { workspace = true }
-aws-credential-types = { workspace = true, features = ["hardcoded-credentials"] }
+#arrow = { workspace = true, features = ["prettyprint"] }
+#aggregator_udfs = { path = "../aggregator_udfs" }
+#aws-config = { workspace = true }
+#aws-credential-types = { workspace = true, features = ["hardcoded-credentials"] }
+##bytes = { workspace = true }
 aws-types = { workspace = true }
-bytes = { workspace = true }
 chrono = { workspace = true }
 color-eyre = { workspace = true }
-cxx = { workspace = true }
-datafusion = { workspace = true }
+#cxx = { workspace = true }
+#datafusion = { workspace = true }
 env_logger = { workspace = true }
-filter_udfs = { path = "../filter_udfs" }
-futures = { workspace = true }
-libc = { workspace = true }
+#filter_udfs = { path = "../filter_udfs" }
+#futures = { workspace = true }
 log = { workspace = true }
-num-format = { workspace = true }
-num-traits = { workspace = true }
-object_store = { workspace = true }
-objectstore_ext = { path = "../objectstore_ext" }
-regex = { workspace = true }
-rust_sketch = { path = "../rust_sketch" }
+libc = { workspace = true }
+#num-format = { workspace = true }
+#num-traits = { workspace = true }
+#object_store = { workspace = true }
+#objectstore_ext = { path = "../objectstore_ext" }
+#regex = { workspace = true }
+#rust_sketch = { path = "../rust_sketch" }
+sleeper_core = { path = "../sleeper_core" }
 tokio = { workspace = true, features = ["full"] }
 url = { workspace = true }
-
-[dev-dependencies]
-tempfile = { workspace = true }
-test-log = { workspace = true }
-tokio-test = { workspace = true }
